backend/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.1
neo4j==5.17.0  # Neo4j Python driver
openai==1.12.0  # OpenAI Python client

# ETL Dependencies
aiohttp==3.9.3  # For async HTTP requests in website scraper
beautifulsoup4==4.12.3  # For HTML parsing
PyMuPDF==1.23.26  # For PDF processing (fitz)
python-docx==1.1.0  # For Word document processing
pandas==2.2.1  # For data manipulation
